Refactor phase types (mostly moving imports)
