+++ x-pack/plugins/index_lifecycle_management/__jest__/client_integration/edit_policy/features/node_allocation/cold_phase.test.ts
@@ -6,7 +6,7 @@
  */
 
 import { act } from 'react-dom/test-utils';
-import { setupEnvironment } from '../../../helpers/setup_environment';
+import { setupEnvironment } from '../../../helpers';
 import { NodeAllocationTestBed, setupColdPhaseNodeAllocation } from './cold_phase.helpers';
 
 describe('<EditPolicy /> node allocation in the cold phase', () => {
