+++ x-pack/plugins/index_lifecycle_management/__jest__/client_integration/edit_policy/features/node_allocation/general_behavior.test.ts
@@ -6,7 +6,7 @@
  */
 
 import { act } from 'react-dom/test-utils';
-import { setupEnvironment } from '../../../helpers/setup_environment';
+import { setupEnvironment } from '../../../helpers';
 import {
   GeneralNodeAllocationTestBed,
   setupGeneralNodeAllocation,
