+++ x-pack/plugins/index_lifecycle_management/__jest__/client_integration/helpers/actions/forcemerge_actions.ts
@@ -7,9 +7,9 @@
 
 import { act } from 'react-dom/test-utils';
 import { TestBed } from '@kbn/test/jest';
+import { Phase } from '../../../../common/types';
 import { createFormToggleAction } from './form_toggle_action';
 import { createFormSetValueAction } from './form_set_value_action';
-import { Phase } from '../types';
 
 const createFormCheckboxAction = (testBed: TestBed, dataTestSubject: string) => async (
   checked: boolean
