+++ x-pack/plugins/index_lifecycle_management/__jest__/client_integration/helpers/actions/index_priority_actions.ts
@@ -6,9 +6,9 @@
  */
 
 import { TestBed } from '@kbn/test/jest';
+import { Phase } from '../../../../common/types';
 import { createFormToggleAction } from './form_toggle_action';
 import { createFormSetValueAction } from './form_set_value_action';
-import { Phase } from '../types';
 
 export const createIndexPriorityActions = (testBed: TestBed, phase: Phase) => {
   const { exists } = testBed;
