+++ x-pack/plugins/index_lifecycle_management/__jest__/client_integration/helpers/actions/searchable_snapshot_actions.ts
@@ -7,7 +7,7 @@
 
 import { act } from 'react-dom/test-utils';
 import { TestBed } from '@kbn/test/jest';
-import { Phase } from '../types';
+import { Phase } from '../../../../common/types';
 import { createFormToggleAction } from './form_toggle_action';
 
 export const createSearchableSnapshotActions = (testBed: TestBed, phase: Phase) => {
