+++ x-pack/plugins/index_lifecycle_management/__jest__/client_integration/helpers/index.ts
@@ -5,6 +5,6 @@
  * 2.0.
  */
 
-export { Phase } from './types';
-
 export * from './actions';
+
+export { setupEnvironment } from './setup_environment';
