--- x-pack/plugins/index_lifecycle_management/__jest__/client_integration/helpers/types.ts
@@ -1,10 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { Phases } from '../../../common/types';
-
-export type Phase = keyof Phases;
