+++ x-pack/plugins/index_lifecycle_management/common/types/policies.ts
@@ -7,7 +7,13 @@
 
 import { Index as IndexInterface } from '../../../index_management/common/types';
 
+export type Phase = keyof Phases;
+
 export type PhaseWithAllocation = 'warm' | 'cold';
+
+export type PhaseWithTiming = keyof Omit<Phases, 'hot'>;
+
+export type PhasesExceptDelete = keyof Omit<Phases, 'delete'>;
 
 export interface SerializedPolicy {
   name: string;
@@ -21,8 +27,6 @@ export interface Phases {
   frozen?: SerializedFrozenPhase;
   delete?: SerializedDeletePhase;
 }
-
-export type PhasesExceptDelete = keyof Omit<Phases, 'delete'>;
 
 export interface PolicyFromES {
   modified_date: string;
