+++ x-pack/plugins/index_lifecycle_management/public/application/sections/edit_policy/components/phases/shared_fields/min_age_field/util.ts
@@ -6,12 +6,9 @@
  */
 
 import { i18n } from '@kbn/i18n';
-import { Phases } from '../../../../../../../../common/types';
+import { PhaseWithTiming } from '../../../../../../../../common/types';
 
-type PhaseWithMinAgeAction = 'warm' | 'cold' | 'delete';
-
-export function getUnitsAriaLabelForPhase(phase: keyof Phases) {
-  // NOTE: Hot phase isn't necessary, because indices begin in the hot phase.
+export function getUnitsAriaLabelForPhase(phase: PhaseWithTiming) {
   switch (phase) {
     case 'warm':
       return i18n.translate(
@@ -29,6 +26,14 @@ export function getUnitsAriaLabelForPhase(phase: keyof Phases) {
         }
       );
 
+    case 'frozen':
+      return i18n.translate(
+        'xpack.indexLifecycleMgmt.editPolicy.phaseFrozen.minimumAgeUnitsAriaLabel',
+        {
+          defaultMessage: 'Units for timing of frozen phase',
+        }
+      );
+
     case 'delete':
       return i18n.translate(
         'xpack.indexLifecycleMgmt.editPolicy.phaseDelete.minimumAgeUnitsAriaLabel',
@@ -38,8 +43,7 @@ export function getUnitsAriaLabelForPhase(phase: keyof Phases) {
       );
   }
 }
-export function getTimingLabelForPhase(phase: PhaseWithMinAgeAction) {
-  // NOTE: Hot phase isn't necessary, because indices begin in the hot phase.
+export function getTimingLabelForPhase(phase: PhaseWithTiming) {
   switch (phase) {
     case 'warm':
       return i18n.translate('xpack.indexLifecycleMgmt.editPolicy.phaseWarm.minimumAgeLabel', {
@@ -49,6 +53,11 @@ export function getTimingLabelForPhase(phase: PhaseWithMinAgeAction) {
     case 'cold':
       return i18n.translate('xpack.indexLifecycleMgmt.editPolicy.phaseCold.minimumAgeLabel', {
         defaultMessage: 'Timing for cold phase',
+      });
+
+    case 'frozen':
+      return i18n.translate('xpack.indexLifecycleMgmt.editPolicy.phaseFrozen.minimumAgeLabel', {
+        defaultMessage: 'Timing for frozen phase',
       });
 
     case 'delete':
